Preliminary approach to queueing profile stream uploads on a separate thread
